Refactor de logeo de eventos Analytics
