+++ app/src/main/java/ar/edu/unq/pdes/myprivateblog/screens/post_detail/PostDetailViewModel.kt
@@ -1,16 +1,14 @@
 package ar.edu.unq.pdes.myprivateblog.screens.post_detail
 
-import android.content.Context
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import ar.edu.unq.pdes.myprivateblog.data.BlogEntry
 import ar.edu.unq.pdes.myprivateblog.data.EntityID
-import ar.edu.unq.pdes.myprivateblog.helper.logEventDeletePost
-import ar.edu.unq.pdes.myprivateblog.helper.logEventEditPost
+import ar.edu.unq.pdes.myprivateblog.logger.AnalyticsLogger
 import ar.edu.unq.pdes.myprivateblog.services.PostService
 import javax.inject.Inject
 
-class PostDetailViewModel @Inject constructor(private val postService: PostService, val context: Context) : ViewModel() {
+class PostDetailViewModel @Inject constructor(private val postService: PostService, val analytics: AnalyticsLogger) : ViewModel() {
 
     enum class State {
         VIEW, DELETED
@@ -25,7 +23,7 @@ class PostDetailViewModel @Inject constructor(private val postService: PostServi
             post.value = it.first
             bodyText.value = it.second
         }
-        logEventEditPost(context)
+        analytics.logEventEditPost()
     }
 
     fun deletePost() {
@@ -33,11 +31,12 @@ class PostDetailViewModel @Inject constructor(private val postService: PostServi
         val disposable = postService.delete(aPost).subscribe {
             state.value = State.DELETED
         }
-        logEventDeletePost(context)
+        analytics.logEventDeletePost()
     }
 
     fun cancelDeletePost() {
         val aPost = post.value ?: return
         val disposable = postService.restore(aPost).subscribe()
+        analytics.logEventCancelDeletePost()
     }
 }
